+++ user-service/src/main/java/com/chatbot/userservice/services/LawyerService.java
@@ -7,14 +7,14 @@ import java.util.Optional;
 
 public interface LawyerService {
     Lawyer createLawyer(Lawyer lawyer);
-    Optional<Lawyer> getLawyerById(Integer id);
+    Optional<Lawyer> getLawyerById(Long id);
     Optional<Lawyer> getLawyerByEmail(String email);
     List<Lawyer> getAllLawyers();
     List<Lawyer> getLawyersBySpecialization(String specialization);
     List<Lawyer> getLawyersByStatus(String status);
     List<Lawyer> getLawyersByMinimumRating(Float minRating);
     Lawyer updateLawyer(Lawyer lawyer);
-    void deleteLawyer(Integer id);
+    void deleteLawyer(Long id);
     boolean authenticateLawyer(String email, String password);
-    void updateLawyerAvailability(Integer lawyerId, String status);
+    void updateLawyerAvailability(Long lawyerId, String status);
 }
