+++ user-service/src/main/java/com/chatbot/userservice/services/NotificationService.java
@@ -6,12 +6,12 @@ import java.util.Optional;
 
 public interface NotificationService {
     Notification createNotification(Notification notification);
-    Optional<Notification> getNotificationById(Integer id);
-    List<Notification> getNotificationsByUserId(Integer userId);
-    List<Notification> getNotificationsByLawyerId(Integer lawyerId);
+    Optional<Notification> getNotificationById(Long id);
+    List<Notification> getNotificationsByUserId(Long userId);
+    List<Notification> getNotificationsByLawyerId(Long lawyerId);
     List<Notification> getNotificationsByType(String type);
-    List<Notification> getUnreadNotifications(Integer userId);
+    List<Notification> getUnreadNotifications(Long userId);
     Notification updateNotification(Notification notification);
-    void deleteNotification(Integer id);
-    void markAsRead(Integer notificationId);
+    void deleteNotification(Long id);
+    void markAsRead(Long notificationId);
 }
