+++ user-service/src/main/java/com/chatbot/userservice/services/UserActivityService.java
@@ -8,11 +8,11 @@ import java.util.Optional;
 
 public interface UserActivityService {
     UserActivity createActivity(UserActivity activity);
-    Optional<UserActivity> getActivityById(Integer id);
-    List<UserActivity> getActivitiesByUserId(Integer userId);
+    Optional<UserActivity> getActivityById(Long id);
+    List<UserActivity> getActivitiesByUserId(Long userId);
     List<UserActivity> getActivitiesByType(String type);
     List<UserActivity> getActivitiesByDateRange(LocalDateTime start, LocalDateTime end);
     UserActivity updateActivity(UserActivity activity);
-    void deleteActivity(Integer id);
-    void logActivity(Integer userId, String activityType, String details);
+    void deleteActivity(Long id);
+    void logActivity(Long userId, String activityType, String details);
 }
