+++ user-service/src/main/java/com/chatbot/userservice/services/UserService.java
@@ -7,12 +7,12 @@ import java.util.Optional;
 
 public interface UserService {
     User createUser(User user);
-    Optional<User> getUserById(Integer id);
+    Optional<User> getUserById(Long id);
     Optional<User> getUserByEmail(String email);
     List<User> getAllUsers();
     List<User> getUsersByStatus(String status);
     User updateUser(User user);
-    void deleteUser(Integer id);
+    void deleteUser(Long id);
     boolean authenticateUser(String email, String password);
-    void changeUserLanguage(Integer userId, String language);
+    void changeUserLanguage(Long userId, String language);
 }
